Remove ValidationExtensions class and its methods
The `ValidationExtensions` class has been completely removed from `ValidationExtensions.cs`. This includes the deletion of the `EnsureUniqueAsync` and `EnsureUniqueAsync<T>` methods, which were responsible for checking the uniqueness of a task and returning an `AppResult` or `AppResult<T>`. This change eliminates any functionality that relied on these methods in the application.